requirements.txt
accelerate==0.30.1
black==24.4.2
datasets==2.19.1
evaluate==0.4.2
matplotlib==3.9.0
nltk==3.8.1
numpy==1.26.4
optuna==3.6.1
pandas==2.2.2
peft==0.11.1
scikit-learn==1.5.0
spacy==3.7.4
streamlit==1.34.0
tokenizers==0.19.1
transformers==4.41.1
xgboost==2.0.3
python-docx==1.1.2
shap==0.45.1
seaborn==0.13.2

